requirements.txt
# Core RL / training stack (pinned from your working env)
unsloth==2025.10.11
unsloth_zoo==2025.10.12
transformers==4.57.1
datasets==4.3.0
peft==0.17.1
trl==0.23.0
vllm==0.11.0
tensorboardX==2.6.4
swanlab==0.6.13
nvitop==1.5.3

# Evaluation / analysis (pinned from your working env)
sacrebleu==2.5.1
sentence-transformers==5.1.2
unbabel-comet==2.2.7
jieba==0.42.1
matplotlib==3.10.7
openpyxl==3.1.5
numpy==1.26.4
tqdm==4.67.1
requests==2.32.5
openai==2.6.1
notebook==7.4.7
